Created bodyscroll
This functions similar to the bigbanner. The main difference being the use of swiper so it's a carousel including the games. Also includes a dynamic rating by the title. Depending on the rating the game has depends the color of the text and border of it. Also includes a bookmark for saving ( not finished yet )