Created gameinfo and gameinfoarray
These will be used in my game file that displays the game the user wants to view with its information. I created separate components, because some of the results inside the game include strings, objects, and arrays. They are basically the same except for the infoarray one mapping through the information due to the result being an array. Also passed in certain styles to change depending on the element in the main document