Created functional search bar
Updated the search bar in the nav so it correctly works. Now it takes the user to the searched page, and using the text they searched results return with games that include or are similar. Searched includes a template similar to grid, but simplified to just title, save, and score