Added game, actualGame, and result
The game state will be used for areas like the gameGrid for calling the game. The getGame function gets the games from rawg. Actual game and result work together for the profile section. getGameId is used to grab the id of the games' bookmark that's clicked. With that click it also sets the result that's then passed into getActualGame which grabs the game using the id