Created profile
This page stores all the users saved games. The way it works is it maps through multiGames. If a games' bookmark is clicked on another page it addas it to the array. Reused most the styles from the gameGrid. Just took away the bookmark and added the remove button inside each game, as well as the remove all element to remove all games quicker