Created functions for pages
First function created is StartAtTop. This function uses location and on pathname changing it sets the window to 0, 0. This guarantees that every time a user visits a new page its at the very top. Also added isActive and revised multiGames. Both have an initial state of an empty array. Once an item is added it's stored in local host and adds on from there. MultiGames takes care of displayed games in the profile, and the isActive state manages each games' bookmark. If clicked will change to a check mark and add game. Revised removeObj and removeAll to remove selected item or all items, as well as change check mark back to bookmark. The changeActive function changes the active state to check if it already includes the item. Then finally on multiGames state changing it updates both arrays in localhost.