Added functional bookmark
The bookmark now uses the isActive state. When clicked the bookmark checks if the games id is already included in the isActive array, if not it adds it to the array as well as adding the game to the profile. After clicking it also changes the bookmark to display a check mark, that using local host stays that way until removed in profile.